requirements.txt
google-api-python-client==2.99.0
google-auth==2.16.0
google-auth-oauthlib==0.6.1
google-auth-httplib2==0.1.0
google-auth-httplib2==0.1.0
pytesseract==0.3.10
Pillow==9.4.0
python-dotenv==0.21.1
requests==2.28.1
